LWCloudProvider 4.2.18.3  - refactor  - change to SDK Style  - organize projects better: separate out test projects
