+++ .gitignore
@@ -255,3 +255,4 @@ IATETerminologyProvider/IATETerminologyProvider/pluginpackage.manifest_LOCAL_178
 /TMBackup/build/de/Sdl.Versioning.resources.dll
 /TMBackup/build/de/Sdl.Core.PluginFramework.resources.dll
 /TMBackup/build/de/Microsoft.Win32.TaskScheduler.resources.dll
+/SDLMTCloud.Provider/Sdl.Community.MTCloud.Provider/Properties/launchSettings.json
