+++ SDLMTCloud.Provider/Sdl.Community.MTCloud.Provider.sln
@@ -1,21 +1,13 @@
 ﻿
 Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio Version 16
-VisualStudioVersion = 16.0.32228.343
+# Visual Studio Version 17
+VisualStudioVersion = 17.1.32421.90
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Sdl.Community.MTCloud.Provider", "Sdl.Community.MTCloud.Provider\Sdl.Community.MTCloud.Provider.csproj", "{C5C7E441-7416-47F3-A4B0-C20AEEC12ABF}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Sdl.Community.MTCloud.Provider", "Sdl.Community.MTCloud.Provider\Sdl.Community.MTCloud.Provider.csproj", "{C5C7E441-7416-47F3-A4B0-C20AEEC12ABF}"
 EndProject
-Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Tests", "Tests", "{9BA60C05-6844-4B2C-B3EC-73C1DAF9A769}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Sdl.Community.MTCloud.Languages.Provider", "SDLMTCloud.Languages.Provider\Sdl.Community.MTCloud.Languages.Provider.csproj", "{9D257236-08A9-4B23-9249-ACA78139172E}"
 EndProject
-Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "IntegrationTests", "IntegrationTests", "{9DCAD356-3AF1-49CB-B84A-A17D7C9E7B2E}"
-EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Sdl.Community.MTCloud.Languages.Provider.IntegrationTests", "..\SDLMTCloud.Languages.Provider\Tests\IntegrationTests\Sdl.Community.MTCloud.Languages.Provider.IntegrationTests\Sdl.Community.MTCloud.Languages.Provider.IntegrationTests.csproj", "{EA39EB5D-0E7E-4EFF-B77A-3A15AD1DE50A}"
-EndProject
-Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "UnitTests", "UnitTests", "{0099060D-B9D7-4D3B-9E6B-26C47C510858}"
-EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Sdl.Community.MTCloud.Provider.UnitTests", "Sdl.Community.MTCloud.Provider\Tests\UnitTests\Sdl.Community.MTCloud.Provider.UnitTests\Sdl.Community.MTCloud.Provider.UnitTests.csproj", "{1D6B3478-0B2D-429B-B7D7-1F45CE7E4B05}"
-EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Sdl.Community.MTCloud.Languages.Provider", "SDLMTCloud.Languages.Provider\Sdl.Community.MTCloud.Languages.Provider.csproj", "{9D257236-08A9-4B23-9249-ACA78139172E}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Tests", "Tests\Tests.csproj", "{8E43C897-DA25-4C4E-BA42-46E7D26FC9ED}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -27,27 +19,17 @@ Global
 		{C5C7E441-7416-47F3-A4B0-C20AEEC12ABF}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{C5C7E441-7416-47F3-A4B0-C20AEEC12ABF}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{C5C7E441-7416-47F3-A4B0-C20AEEC12ABF}.Release|Any CPU.Build.0 = Release|Any CPU
-		{EA39EB5D-0E7E-4EFF-B77A-3A15AD1DE50A}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{EA39EB5D-0E7E-4EFF-B77A-3A15AD1DE50A}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{EA39EB5D-0E7E-4EFF-B77A-3A15AD1DE50A}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{EA39EB5D-0E7E-4EFF-B77A-3A15AD1DE50A}.Release|Any CPU.Build.0 = Release|Any CPU
-		{1D6B3478-0B2D-429B-B7D7-1F45CE7E4B05}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{1D6B3478-0B2D-429B-B7D7-1F45CE7E4B05}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{1D6B3478-0B2D-429B-B7D7-1F45CE7E4B05}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{1D6B3478-0B2D-429B-B7D7-1F45CE7E4B05}.Release|Any CPU.Build.0 = Release|Any CPU
 		{9D257236-08A9-4B23-9249-ACA78139172E}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{9D257236-08A9-4B23-9249-ACA78139172E}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{9D257236-08A9-4B23-9249-ACA78139172E}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{9D257236-08A9-4B23-9249-ACA78139172E}.Release|Any CPU.Build.0 = Release|Any CPU
+		{8E43C897-DA25-4C4E-BA42-46E7D26FC9ED}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{8E43C897-DA25-4C4E-BA42-46E7D26FC9ED}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{8E43C897-DA25-4C4E-BA42-46E7D26FC9ED}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{8E43C897-DA25-4C4E-BA42-46E7D26FC9ED}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
-	EndGlobalSection
-	GlobalSection(NestedProjects) = preSolution
-		{9DCAD356-3AF1-49CB-B84A-A17D7C9E7B2E} = {9BA60C05-6844-4B2C-B3EC-73C1DAF9A769}
-		{EA39EB5D-0E7E-4EFF-B77A-3A15AD1DE50A} = {9DCAD356-3AF1-49CB-B84A-A17D7C9E7B2E}
-		{0099060D-B9D7-4D3B-9E6B-26C47C510858} = {9BA60C05-6844-4B2C-B3EC-73C1DAF9A769}
-		{1D6B3478-0B2D-429B-B7D7-1F45CE7E4B05} = {0099060D-B9D7-4D3B-9E6B-26C47C510858}
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
 		SolutionGuid = {9701082D-D682-4276-BECF-B295729EB24F}
