+++ SDLMTCloud.Provider/Sdl.Community.MTCloud.Provider/Properties/AssemblyInfo.cs
@@ -7,12 +7,8 @@ using System.Runtime.InteropServices;
 // set of attributes. Change these attribute values to modify the information
 // associated with an assembly.
 [assembly: AssemblyTitle("SDL Machine Translation Cloud provider")]
-[assembly: AssemblyDescription("")]
-[assembly: AssemblyConfiguration("")]
 [assembly: AssemblyProduct("SDL Machine Translation Cloud provider")]
 [assembly: AssemblyCopyright("Copyright © SDL 2018")]
-[assembly: AssemblyTrademark("")]
-[assembly: AssemblyCulture("")]
 
 // Setting ComVisible to false makes the types in this assembly not visible 
 // to COM components.  If you need to access a type in this assembly from 
@@ -33,4 +29,4 @@ using System.Runtime.InteropServices;
 // by using the '*' as shown below:
 // [assembly: AssemblyVersion("1.0.*")]
 [assembly: AssemblyVersion("4.2.18.0")]
-[assembly: AssemblyFileVersion("4.2.18.2")]
+[assembly: AssemblyFileVersion("4.2.18.3")]
