--- SDLMTCloud.Provider/Sdl.Community.MTCloud.Provider/packages.config
@@ -1,9 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8"?>
-<packages>
-  <package id="MahApps.Metro" version="1.5.0" targetFramework="net47" />
-  <package id="Microsoft.Office.Interop.Excel" version="15.0.4795.1000" targetFramework="net47" />
-  <package id="Sdl.Community.Toolkit.Core" version="1.8.8" targetFramework="net47" />
-  <package id="Sdl.Community.Toolkit.LanguagePlatform" version="1.9.0" targetFramework="net47" />
-  <package id="Sdl.Core.PluginFramework" version="1.8.0" targetFramework="net461" />
-  <package id="Sdl.Core.PluginFramework.Build" version="15.0.2" targetFramework="net461" />
-</packages>
