+++ SDLMTCloud.Provider/Sdl.Community.MTCloud.Provider/pluginpackage.manifest.xml
@@ -1,7 +1,7 @@
 ﻿<?xml version="1.0" encoding="utf-8"?>
 <PluginPackage xmlns="http://www.sdl.com/Plugins/PluginPackage/1.0">
 	<PlugInName>Language Weaver</PlugInName>
-	<Version>4.2.18.2</Version>
+	<Version>4.2.18.3</Version>
 	<Description>Language Weaver Provider</Description>
 	<Author>Trados AppStore Team</Author>
 	<Include>
