+++ SDLMTCloud.Provider/Tests/Tests.csproj
@@ -1,0 +1,29 @@
+﻿<?xml version="1.0" encoding="utf-8"?>
+<Project Sdk="Microsoft.NET.Sdk">
+	<PropertyGroup>
+	 <TargetFramework>net48</TargetFramework>
+	 <GenerateAssemblyInfo>false</GenerateAssemblyInfo>
+	</PropertyGroup>
+	<ItemGroup>
+	  <PackageReference Include="NSubstitute" Version="4.3.0" />
+	  <PackageReference Include="xunit" Version="2.4.1" />
+	  <PackageReference Include="xunit.assert" Version="2.4.1" />
+	  <PackageReference Include="xunit.extensibility.core" Version="2.4.1" />
+	</ItemGroup>
+	<ItemGroup>
+	  <ProjectReference Include="..\Sdl.Community.MTCloud.Provider\Sdl.Community.MTCloud.Provider.csproj" />
+	</ItemGroup>
+  <ItemGroup>
+    <Reference Include="Sdl.LanguagePlatform.TranslationMemoryApi">
+      <HintPath>$(MSBuildProgramFiles32)\SDL\SDL Trados Studio\Studio16\Sdl.LanguagePlatform.TranslationMemoryApi.dll</HintPath>
+    </Reference>
+    <Reference Include="System" />
+    <Reference Include="System.Core" />
+    <Reference Include="System.Xml.Linq" />
+    <Reference Include="System.Data.DataSetExtensions" />
+    <Reference Include="Microsoft.CSharp" />
+    <Reference Include="System.Data" />
+    <Reference Include="System.Net.Http" />
+    <Reference Include="System.Xml" />
+  </ItemGroup>
+</Project>
